requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
click==8.0.3
fastapi==0.73.0
h11==0.13.0
idna==3.3
motor==2.5.1
pycodestyle==2.8.0
pydantic==1.9.0
pymongo==3.12.3
python-dotenv==0.19.2
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.17.1
